split the error_code.go file into multiple files
